Fix(ui) LAI2-8818: loading spinner for unavailable <iframe> (#84)
Implements a simple check which calls 'fetch()' on the iframe's
target URL and renders a loading spinnner if it does not return
HTTP 200.

NOTE: Because of browser security measures, we can only check
URLs with the same hostname (CORS). Therefore, this will only
work for the Streamlit/StaticWebFrontend components, and only
in the cloud (where they are served using the same Ingress URL).
Locally, we run these servers with different localhost ports,
which the browser blocks.